[moe training] rename moe_training -> fp8_grouped_mm
