+++ benchmarks/prototype/moe_training/fp8_rowwise/bench_triton_fp8_rowwise_3d_transpose_rhs.py
@@ -14,11 +14,11 @@ from tabulate import tabulate
 from tqdm import tqdm
 from triton.testing import do_bench
 
-from torchao.prototype.moe_training.kernels.float8_rowwise import (
+from torchao.prototype.fp8_grouped_mm.kernels.float8_rowwise import (
     triton_fp8_rowwise_3d_transpose_rhs,
     triton_fp8_rowwise_3d_transpose_rhs_fused_reduction,
 )
-from torchao.prototype.moe_training.utils import (
+from torchao.prototype.fp8_grouped_mm.utils import (
     torch_to_3d_rowwise_float8_transpose_rhs,
 )
