+++ benchmarks/prototype/moe_training/mxfp8/bench_mx_block_rearrange_2d_K_groups.py
@@ -14,7 +14,7 @@ from tabulate import tabulate
 from tqdm import tqdm
 
 from benchmarks.utils import benchmark_cuda_function_in_microseconds
-from torchao.prototype.moe_training.utils import generate_jagged_offs
+from torchao.prototype.fp8_grouped_mm.utils import generate_jagged_offs
 from torchao.prototype.mx_formats.grouped_mm.kernels import (
     torch_to_blocked_2d_K_groups,
     triton_mx_block_rearrange_2d_K_groups,
