+++ torchao/prototype/fp8_grouped_mm/kernels/__init__.py
@@ -1,9 +1,9 @@
-from torchao.prototype.moe_training.kernels.float8_rowwise import (
+from torchao.prototype.fp8_grouped_mm.kernels.float8_rowwise import (
     triton_fp8_rowwise_3d_transpose_rhs as triton_fp8_rowwise_3d_transpose_rhs,
 )
-from torchao.prototype.moe_training.kernels.jagged_float8_scales import (
+from torchao.prototype.fp8_grouped_mm.kernels.jagged_float8_scales import (
     triton_fp8_per_group_colwise_scales as triton_fp8_per_group_colwise_scales,
 )
-from torchao.prototype.moe_training.kernels.jagged_float8_scales import (
+from torchao.prototype.fp8_grouped_mm.kernels.jagged_float8_scales import (
     triton_fp8_per_group_rowwise_scales as triton_fp8_per_group_rowwise_scales,
 )
