+++ torchao/prototype/mx_formats/ep/a2a_dispatch.py
@@ -9,7 +9,7 @@ import torch.distributed as dist
 from torch.distributed._functional_collectives import all_to_all_single
 from torch.distributed.distributed_c10d import _resolve_process_group
 
-from torchao.prototype.moe_training.utils import conditional_nostrict_trace
+from torchao.prototype.fp8_grouped_mm.utils import conditional_nostrict_trace
 from torchao.prototype.mx_formats.config import ScaleCalculationMode
 from torchao.prototype.mx_formats.kernels import triton_to_mxfp8_dim0
 from torchao.prototype.mx_formats.mx_tensor import MXTensor
